start implementing backend API, and some bikeshedding

+++ internal/backend/dbInterface/dbInterface.go
@@ -5,10 +5,11 @@
 -> generate password
 */
 
-package backend
+package dbInterface
 
 import (
 	"database/sql"
+	"errors"
 	"fmt"
 	"passwordManager/internal/userType"
 
@@ -34,6 +35,9 @@ func OpenDb() error {
 
 func InsertUser(username string, salt []byte, masterKeyHash []byte) (string, error) {
 	//returns the username of the user created, or possible error
+	if len(username) == 0 {
+		return "", errors.New("db: username given to add user has length 0")
+	}
 
 	statement, err := db.Prepare("INSERT INTO users (username, salt, key_hash) VALUES (?, ?, ?)")
 
@@ -55,6 +59,9 @@ func InsertUser(username string, salt []byte, masterKeyHash []byte) (string, err
 
 func DeleteUser(username string, uid int64) (string, error) {
 	//returns the username of the deleted user, or possible error
+	if len(username) == 0 {
+		return "", errors.New("db: username given to delete user has length 0")
+	}
 
 	statement, err := db.Prepare("DELETE FROM users WHERE id = ? AND username = ?")
 	if err != nil {
@@ -75,6 +82,9 @@ func DeleteUser(username string, uid int64) (string, error) {
 
 func FetchUser(username string) (userType.User, error) {
 	//returns the user information, error otherwise
+	if len(username) == 0 {
+		return userType.User{}, errors.New("db: username given to fetch user has length 0")
+	}
 
 	row := db.QueryRow("SELECT id, username, salt, key_hash FROM users WHERE username = ?", username)
 
@@ -90,6 +100,9 @@ func FetchUser(username string) (userType.User, error) {
 
 func FetchPassword(uid int64, accountName string) ([]byte, error) {
 	//returns the hashed password corresponding to the account, or possible error
+	if len(accountName) == 0 {
+		return []byte{}, errors.New("db: account name given to fetch password has length 0")
+	}
 
 	row := db.QueryRow("SELECT encrypted_data FROM entries WHERE user_id = ? AND name = ?", uid, accountName)
 
@@ -104,6 +117,9 @@ func FetchPassword(uid int64, accountName string) ([]byte, error) {
 
 func InsertPassword(uid int64, accountName string, hashedPassword []byte) (string, error) {
 	//returns the name of the account for which a password was added, or a possible error
+	if len(accountName) == 0 {
+		return "", errors.New("db: username given to insert password has length 0")
+	}
 
 	statement, err := db.Prepare("INSERT INTO entries (user_id, name, encrypted_data) VALUES (?, ?, ?)")
 	if err != nil {
